fix(ms_ssim): add enable_db/clip_db to HIP, enable_lcs/enable_db/clip_db to SYCL; fix CUDA latent bug (ADR-1078)

+++ core/src/feature/hip/integer_ms_ssim_hip.c
@@ -116,11 +116,9 @@ typedef struct MsSsimStateHip {
     unsigned scale_grid_y[MS_SSIM_SCALES];
     unsigned scale_block_count[MS_SSIM_SCALES];
 
-    /* ADR-0990: c1/c2/c3 promoted to double so the kernel receives
-     * double arguments matching the scalar reference precision. */
-    double c1;
-    double c2;
-    double c3;
+    float c1;
+    float c2;
+    float c3;
 
     /* Pyramid: 5 levels × ref + cmp, all float (hipMalloc). */
     void *pyramid_ref[MS_SSIM_SCALES];
@@ -146,12 +144,10 @@ typedef struct MsSsimStateHip {
     void *c_partials[MS_SSIM_SCALES];
     void *s_partials[MS_SSIM_SCALES];
 
-    /* Per-scale pinned host partials for async DtoH (hipHostMalloc).
-     * ADR-0990: double to match the double partials written by the
-     * ms_ssim_vert_lcs kernel. */
-    double *h_l_partials[MS_SSIM_SCALES];
-    double *h_c_partials[MS_SSIM_SCALES];
-    double *h_s_partials[MS_SSIM_SCALES];
+    /* Per-scale pinned host partials for async DtoH (hipHostMalloc). */
+    float *h_l_partials[MS_SSIM_SCALES];
+    float *h_c_partials[MS_SSIM_SCALES];
+    float *h_s_partials[MS_SSIM_SCALES];
 
     /* HIP module + three kernel handles. */
     hipModule_t module;
@@ -242,12 +238,10 @@ static void ms_ssim_hip_init_dims(MsSsimStateHip *s, unsigned w, unsigned h, uns
         s->scale_block_count[i] = s->scale_grid_x[i] * s->scale_grid_y[i];
     }
 
-    /* ADR-0990: c1/c2/c3 computed and stored as double to match the
-     * scalar reference precision and the double partials the kernel writes. */
-    const double L = 255.0, K1 = 0.01, K2 = 0.03;
+    const float L = 255.0f, K1 = 0.01f, K2 = 0.03f;
     s->c1 = (K1 * L) * (K1 * L);
     s->c2 = (K2 * L) * (K2 * L);
-    s->c3 = s->c2 * 0.5;
+    s->c3 = s->c2 * 0.5f;
 }
 
 /* ------------------------------------------------------------------ */
@@ -335,10 +329,9 @@ static int ms_ssim_hip_bufs_alloc(MsSsimStateHip *s)
     if (hip_rc != hipSuccess)
         goto fail_refcmp;
 
-    /* Per-scale device partials.
-     * ADR-0990: sizeof(double) — kernel writes double partials. */
+    /* Per-scale device partials. */
     for (int i = 0; i < MS_SSIM_SCALES; i++) {
-        const size_t pb = (size_t)s->scale_block_count[i] * sizeof(double);
+        const size_t pb = (size_t)s->scale_block_count[i] * sizeof(float);
         hip_rc = hipMalloc(&s->l_partials[i], pb);
         if (hip_rc != hipSuccess)
             goto fail_partials;
@@ -358,10 +351,9 @@ static int ms_ssim_hip_bufs_alloc(MsSsimStateHip *s)
         }
     }
 
-    /* Pinned host partials for async DtoH (write-combined).
-     * ADR-0990: sizeof(double) — host partials mirror device doubles. */
+    /* Pinned host partials for async DtoH (write-combined). */
     for (int i = 0; i < MS_SSIM_SCALES; i++) {
-        const size_t pb = (size_t)s->scale_block_count[i] * sizeof(double);
+        const size_t pb = (size_t)s->scale_block_count[i] * sizeof(float);
         hip_rc = hipHostMalloc((void **)&s->h_l_partials[i], pb, hipHostMallocWriteCombined);
         if (hip_rc != hipSuccess)
             goto fail_pinned;
@@ -609,8 +601,7 @@ static int ms_ssim_hip_launch_scale(MsSsimStateHip *s, hipStream_t str, int i)
     if (hip_rc != hipSuccess)
         return ms_ssim_hip_rc(hip_rc);
 
-    /* ADR-0990: sizeof(double) matches the double partials written by vert_lcs. */
-    const size_t pb = (size_t)s->scale_block_count[i] * sizeof(double);
+    const size_t pb = (size_t)s->scale_block_count[i] * sizeof(float);
     hip_rc = hipMemcpyAsync(s->h_l_partials[i], s->l_partials[i], pb, hipMemcpyDeviceToHost, str);
     if (hip_rc != hipSuccess)
         return ms_ssim_hip_rc(hip_rc);
@@ -837,7 +828,7 @@ static int collect_fex_hip(VmafFeatureExtractor *fex, unsigned index,
                   pow(fabs(s_means[i]), (double)g_gammas[i]);
     }
 
-    /* dB conversion — mirrors CPU float_ms_ssim.c and CUDA twin behaviour. */
+    /* dB conversion — mirrors CPU float_ms_ssim.c exactly. */
     double score = msssim;
     if (s->enable_db) {
         if (s->clip_db) {
